hypeandplay/requirement.txt
asgiref==3.5.2
attrs==22.1.0
cffi==1.15.1
cryptography==38.0.4
Django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2
inflection==0.5.1
jsonschema==4.16.0
Pillow==9.2.0
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.18.1
pytz==2022.5
PyYAML==6.0
sqlparse==0.4.3
tzdata==2022.5
uritemplate==4.1.1
whitenoise==6.2.0
